all reorganize from the grbl folder into the robot folder. SVM and Aruco are working, still need to integrate mask_rcnn into this whole thing. Ready to be connected up and tested at this point. I want to see if reducing the frequency at which I check status will change anything as to the lag I am experiencing.
As for non software related issues, I still need to rebuild (print or cut) a new matrix which has less slack and is closed on the ends. I think this will improve consistency and reduce space taken up by the matrix, maybe making the end robot smaller. The printer is acting up so I cannot get my next revision of parts printed right now but I am learning how to resolve this.

These are the main limitations which are holding me from doing a starting print. Really not that much. I have come such a long way and am proud of myself